move DrawRandomParams to a separate file
